Add customization options for letters j/J and v/V
Some people (notably the author of LLPSI) prefer to use i/I even if
it's a consonant. The same goes for u/U.

Replacing those characters in the canonical form that is shown in the
echo area is one solution. Another would be to just replace them for
determining matches, like it's done for macrons. For now, go with the
former solution.