Add a function that highlights unknown words
The match data mentioned in commit 0ffe92a is used to decide whether
to underline words in red (for entirely unknown words) and or in
orange (for near matches).

For now, only overlays for the word at point are ever changed. This
might be confusing, so the function is not yet called anywhere. (To
test it regardless, replace grammaticus-lookup in the definition of
the minor mode with grammaticus--check).