In result list, replace boolean flag by correction
This means that entries for exact matches have nil as their car, as
there's nothing to correct in that case. While it may seem a little
strange, it's exactly what is required for having both a correction
mechanism (in the future) and highlighting that differentiates
between exact and near matches.