Correctly handle typewriter apostrophes
In Latin, unlike in English, those should only ever occur as quotation
marks, never as part of a word. Consequently, introduce a syntax table
that inherits from text-mode-syntax-table (except for the apostrophe's
entry, of course) and ensure that all functions work with respect to
that syntax table (i.e., no longer search for regular expressions).

The latter change is unequivocally the right choice, but it remains to
be seen whether the same can be said about introducing a custom syntax
table instead of using the one associated with the buffer. For now, go
with what works out of the box.